Remove re-linking programs approach
Remove re-linking programs as redundant.
It costs (cheaper to link once than twice) and is not that common GL API
usage pattern (subjective).
Initial idea was to enable work on optimization what would reduce number
of attrib setup calls in case that VAO is not available (#9433).

As such optimization is not implemented, we can remove re-linking.

Related to closed PR #9433 and PR #11583.